add to feature branch for testing

+++ code/headers/hsv_s.hpp
@@ -8,8 +8,8 @@ namespace r2d2::led_strip {
      *  Struct which contains the hue, saturation and value
      */
     struct hsv_s {
-        uint8_t hue;
-	    uint8_t saturation;
+        uint16_t hue;
+	uint8_t saturation;
         uint8_t value;
     };
 } // namespace r2d2::led_strip
